Fix build with CFLAGS -Wextra warning reports (#3428)

+++ pjlib/src/pj/ip_helper_generic.c
@@ -511,7 +511,7 @@ static pj_status_t get_ipv6_deprecated(unsigned *count, pj_sockaddr addr[])
         struct nlmsghdr *nlmsg_ptr = (struct nlmsghdr *) read_buffer;
         int nlmsg_len = read_size;
 
-        if (nlmsg_len < sizeof (struct nlmsghdr))
+        if (nlmsg_len < (int)sizeof(struct nlmsghdr))
             return PJ_ETOOSMALL;
 
         if (nlmsg_ptr->nlmsg_type == NLMSG_DONE)
@@ -587,7 +587,7 @@ PJ_DEF(pj_status_t) pj_enum_ip_interface2( const pj_enum_ip_option *opt,
         pj_sockaddr deprecatedAddrs[*p_cnt];
         unsigned deprecatedCount = *p_cnt;
         unsigned cnt = 0;
-        int i;
+        unsigned i;
         pj_status_t status;
 
         status = get_ipv6_deprecated(&deprecatedCount, deprecatedAddrs);
@@ -599,8 +599,8 @@ PJ_DEF(pj_status_t) pj_enum_ip_interface2( const pj_enum_ip_option *opt,
             return status;
 
         for (i = 0; i < *p_cnt; ++i) {
-            int j;
-            
+            unsigned j;
+
             ifs[cnt++] = addrs[i];
 
             if (addrs[i].addr.sa_family != pj_AF_INET6())
